+++ src/ros/rover/arm/arm_kinematics/arm_kinematics/src/joint_map/transmission_joint_map.cpp
@@ -29,9 +29,9 @@ TransmissionJointMap::TransmissionJointMap(
         "TransmissionJointMap: gather index out of range for the declared input_count");
     }
   }
-  tx_inputs_.assign(gather_indices_.size(), 0.0F);
-  tx_outputs_.assign(output_count_, 0.0F);
-  scratch_.assign(compute_->scratch_size(), 0.0F);
+  tx_inputs_.assign(gather_indices_.size(), 0.0);
+  tx_outputs_.assign(output_count_, 0.0);
+  scratch_.assign(compute_->scratch_size(), 0.0);
 }
 
 TransmissionJointMap::TransmissionJointMap(const TransmissionJointMap & other)
@@ -39,9 +39,9 @@ TransmissionJointMap::TransmissionJointMap(const TransmissionJointMap & other)
     gather_indices_(other.gather_indices_),
     input_count_(other.input_count_),
     output_count_(other.output_count_),
-    tx_inputs_(other.tx_inputs_.size(), 0.0F),
-    tx_outputs_(other.tx_outputs_.size(), 0.0F),
-    scratch_(other.scratch_.size(), 0.0F)
+    tx_inputs_(other.tx_inputs_.size(), 0.0),
+    tx_outputs_(other.tx_outputs_.size(), 0.0),
+    scratch_(other.scratch_.size(), 0.0)
 {
 }
 
@@ -54,15 +54,15 @@ TransmissionJointMap & TransmissionJointMap::operator=(const TransmissionJointMa
   gather_indices_ = other.gather_indices_;
   input_count_ = other.input_count_;
   output_count_ = other.output_count_;
-  tx_inputs_.assign(other.tx_inputs_.size(), 0.0F);
-  tx_outputs_.assign(other.tx_outputs_.size(), 0.0F);
-  scratch_.assign(other.scratch_.size(), 0.0F);
+  tx_inputs_.assign(other.tx_inputs_.size(), 0.0);
+  tx_outputs_.assign(other.tx_outputs_.size(), 0.0);
+  scratch_.assign(other.scratch_.size(), 0.0);
   return *this;
 }
 
 void TransmissionJointMap::map(
-  const span<const float> inputs,
-  const span<float> outputs) const
+  const span<const double> inputs,
+  const span<double> outputs) const
 {
   if (inputs.size() != input_count_) {
     throw std::invalid_argument("TransmissionJointMap::map() received inputs with the wrong size");
@@ -73,20 +73,20 @@ void TransmissionJointMap::map(
   assert(compute_ && "TransmissionJointMap::map() called on a default-constructed instance");
 
   // Gather: pull each transmission input from the joint map's input vector.
-  const float * __restrict__ in = inputs.data_;
+  const double * __restrict__ in = inputs.data_;
   for (size_t i = 0; i < gather_indices_.size(); ++i) {
     tx_inputs_[i] = in[gather_indices_[i]];
   }
 
   // Compute.
   compute_->compute(
-    span<const float>(tx_inputs_.data(), tx_inputs_.size()),
-    span<float>(tx_outputs_.data(), tx_outputs_.size()),
-    span<float>(scratch_.data(), scratch_.size()));
+    span<const double>(tx_inputs_.data(), tx_inputs_.size()),
+    span<double>(tx_outputs_.data(), tx_outputs_.size()),
+    span<double>(scratch_.data(), scratch_.size()));
 
   // Copy outputs densely. The composite (if any) handles further scatter into final positions
   // or downstream scratch slots.
-  float * __restrict__ out = outputs.data_;
+  double * __restrict__ out = outputs.data_;
   for (size_t i = 0; i < output_count_; ++i) {
     out[i] = tx_outputs_[i];
   }
